implement sass; cleanup game component
