+++ app/page.js
@@ -1,6 +1,8 @@
+'use client';
+
 import Image from "next/image";
 import styles from "./page.module.css";
-import Game from '../components/game.js';
+import Game from '../components/game/game.js';
 
 export default function Home() {
   return (
