+++ components/game/game.module.scss
@@ -1,0 +1,46 @@
+$primary-color: #2600ff;
+
+:export {
+  primaryColor: $primary-color;
+}
+
+.shuffleBtn {
+  background-color: $primary-color;
+  color: white;
+  border: none;
+  border-radius: 15px;
+  padding: 1em;
+  margin: 1em;
+  cursor: pointer;
+  font-size: 1.5em;
+}
+
+.intro {
+  background-color: $primary-color;
+  color: white;
+  margin: 1em;
+  padding: 1em;
+  line-height: 1.5;
+  border-radius: 15px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.fade5 { animation: fadeIn 5s; }
+.fade10 { animation: fadeIn 10s; }
+
+@keyframes fadeIn {
+  0% { opacity: 0; }
+  100% { opacity: 1; }
+}
+
+.showShuffle {
+  animation: cssAnimation 0s 5s forwards;
+  visibility: hidden;
+}
+
+@keyframes cssAnimation {
+  to { visibility: visible; }
+}
